CLN: refactor MultiImageReader to BaseMultiImageReader, now an ABCHasTraits class
